Initial listener
Currently doesn't emit events in any way, simply stores the latest of
them into the client's state.
As a side note, it sure is hard to write code against a network device
during the middle of a deep-freeze with rolling blackouts. Looks like I
picked the wrong week to quit sniffing glue.