Refactor to flatten observations
The UDP API should never actually HAVE multiple observations in the
arrays, allegedly, however we want to make sure we can cope if it does.
Because of that we'll emit via PubSub each observation, in timestamp
order, but then only store the most recent one into the state.
This also standardizes the events emitted via pubsub so that pattern
matching events is easier.