Update deps, remove .tool-versions
Needed updated deps to bring in updated ExDoc

We don't really need a .tool-versions, we can just use whatever is
user default, as long as it's greater than the minimum in mix.exs.
This saves us a bunch of problems with older erlang versions not
building correctly, as well as having to constantly keep the asdf
versions updated.